refactor : etc 패키지 constants로 네이밍 변경
