+++ src/main/kotlin/racingCar/constants/Message.kt
@@ -1,4 +1,4 @@
-package racingCar.etc
+package racingCar.constants
 
 /**
  * 출력만 되는 메세지들을 모아둔 Object class
